+++ R/go.R
@@ -169,7 +169,7 @@ fetch_go_species <- function(on_error = c("stop", "warn")) {
 #' @noRd
 fetch_go_genes_go <- function(species, use_cache, on_error = "stop") {
   # Binding variables from non-standard evaluation locally
-  gene_synonym <- db_object_synonym <- symbol <- NULL
+  gene_id <- db_object_synonym <- symbol <- NULL
   db_id <- go_term <- NULL
 
   url <- get_go_annotation_url()
@@ -180,8 +180,8 @@ fetch_go_genes_go <- function(species, use_cache, on_error = "stop") {
   lpath <- cached_url_path(stringr::str_glue("go_gaf_{species}"), gaf_file, use_cache)
   readr::read_tsv(lpath, comment = "!", quote = "", col_names = GAF_COLUMNS,
                   col_types = GAF_TYPES) |>
-    dplyr::mutate(gene_synonym = stringr::str_remove(db_object_synonym, "\\|.*$")) |>
-    dplyr::select(gene_symbol = symbol, gene_synonym, db_id, term_id = go_term) |>
+    dplyr::mutate(gene_id = stringr::str_remove(db_object_synonym, "\\|.*$")) |>
+    dplyr::select(gene_symbol = symbol, gene_id, db_id, term_id = go_term) |>
     dplyr::distinct()
 }
 
@@ -196,7 +196,7 @@ fetch_go_genes_go <- function(species, use_cache, on_error = "stop") {
 #'   file (DB Object Symbol) is returned as \code{gene_symbol}, but, depending
 #'   on the \code{species} argument it can contain other entities, e.g. RNA or
 #'   protein complex names. Similarly, the eleventh column of the GAF file (DB
-#'   Object Synonym) is returned as \code{gene_synonym}. It is up to the user to
+#'   Object Synonym) is returned as \code{gene_id}. It is up to the user to
 #'   select the appropriate database.
 #'
 #' @param species Species designation. Examples are \file{goa_human} for human,
@@ -297,7 +297,7 @@ fetch_go_from_bm <- function(mart, use_cache = TRUE) {
 #'   Symbol) is returned as \code{gene_symbol}, but, depending on the
 #'   \code{species} argument it can contain other entities, e.g. RNA or protein
 #'   complex names. Similarly, the eleventh column of the GAF file (DB Object
-#'   Synonym) is returned as \code{gene_synonym}. It is up to the user to select
+#'   Synonym) is returned as \code{gene_id}. It is up to the user to select
 #'   the appropriate database.
 #'
 #'   Alternatively, if \code{mart} is provided, mapping will be downloaded from